[SPARK-47157][SQL] Refactor file listing with ScanFileListing interface
### What changes were proposed in this pull request?

In this pull request, we've introduce the `ScanFileListing` trait and its implementation, the `GenericScanFileListing` class, to encapsulate and streamline the handling of file listing results. This new abstraction enhances modularity and facilitates more flexible management of file listings within the system.

### Why are the changes needed?

The introduction of these constructs is crucial for defining a standardized API for file listing operations,  regardless of the underlying representation that's used to represent files and partitions. By improving the modularity of the code we enable future improvements that can prove to be beneficial both for runtime and memory improvements.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?

This is just a refactoring, not a new behavior, so existing tests would suffice.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #45224 from costas-db/refactorFileListing.

Lead-authored-by: Costas Zarifis <costas.zarifis@databricks.com>
Co-authored-by: Shoumik Palkar <shoumik.palkar@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>