[SPARK-47202][PYTHON] Fix typo breaking datetimes with tzinfo
### What changes were proposed in this pull request?
This PR fixes a bug caused due to a typo.

### Why are the changes needed?
This bug is preventing users from having datetime.datetime objects with tzinfo when using the `TimestampType`

### Does this PR introduce _any_ user-facing change?
No, just a bug fix.

### How was this patch tested?
There ought to be CI that lints code and catches these simple errors at the time of opening the PR.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #45301 from arzavj/SPARK-47202.

Authored-by: Arzav Jain <arzavj@users.noreply.github.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>