[SPARK-47231][CORE][TESTS] FakeTask should reference its TaskMetrics to avoid TaskMetrics accumulators being GCed before stage completion
### What changes were proposed in this pull request?

I was writing some tests in Spark Core and found this log from time to time
```
INFO AccumulatorContext: Attempted to access garbage collected accumulator 2
```

The reason is, unit tests create `FakeTask` directly without a `Stage`, and the `FakeTask` creates a temporary `TaskMetrics` object then keeps the serialized bytes. This means the `TaskMetrics` accumulators are not referenced by anyone and will be GCed.

This PR fixes this issue by referencing the temporary `TaskMetrics` in `FakeTask`, to avoid the confusing logs in the unit tests.

### Why are the changes needed?

avoid confusing logs when running unit tests

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

N/A, test-only change

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #45336 from cloud-fan/gc.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>