[SPARK-46961][SS] Using ProcessorContext to store and retrieve handle
Setting the processorHandle as a part of the statefulProcessor, so that the user doesn't have to explicitly keep track of it, and can instead simply call `getStatefulProcessorHandle`

This enhances the usability of the State API

Yes, this is an API change. This enhances usability of the StatefulProcessorHandle and the TransformWithState operator.

Existing unit tests are sufficient

No

Closes #45359 from ericm-db/handle-context.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>